requirements.txt
langchain
langchain-core

langchain-openai
openai

langchain-anthropic

langchain-google-genai
google-generativeai

langchain-huggingface
transformers
huggingface-hub 
huggingface_hub[hf_xet]
hf_xet
sentence-transformers

python-dotenv

numpy 
scikit-learn
pytorch